requirements.txt
 Django==1.6.5
MySQL-python==1.2.5
djangorestframework==2.3.13
wsgiref==0.1.2
